First prototype with on the fly UI updates.
The idea is that even during scanning, it should be possible to explore
partially indexed directories. However, in the current implementation each
scanned file generates a UI update via runLater { } which effectively kills
the UI. Some sort of throttling would have to be introduced.

As it turns out, scanning even large folders is relatively fast, so we'll
just make the user wait (existing software I'm aware of does the same
thing).